backend/requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
requests==2.32.3
markdown2==2.5.0
pydantic==2.9.2
openai==0.28.0
google-generativeai==0.7.2
